ML projects/food_classifier_with_CNN/requirements.txt
Keras==2.3.1
matplotlib==3.1.1
google_images_download==2.8.0
keras_lr_finder==0.1
tensorflow_gpu==2.0.0
numpy==1.17.3
Pillow==6.2.1
scikit_learn==0.21.3
tensorflow==2.0.0
